+++ src/index.ts
@@ -11,7 +11,13 @@ export { Sandbox } from "./sandbox.js";
 export type { WaitOptions } from "./sandbox.js";
 
 export { SandboxConnection, SandboxFiles } from "./sandboxd.js";
-export type { ReadFileOptions, WriteFileOptions, WriteFileResult } from "./sandboxd.js";
+export type {
+  FileEntry,
+  ListFilesOptions,
+  ReadFileOptions,
+  WriteFileOptions,
+  WriteFileResult,
+} from "./sandboxd.js";
 
 export type {
   ListSandboxesParams,